Relocate dropzone inline js to separate file
Fixes #21651
