+++ core/http_api.php
@@ -233,13 +233,6 @@ function http_security_headers() {
 			http_csp_add( 'img-src', 'ajax.googleapis.com' );
 		}
 
-		# Relaxing policy for view issue page to allow inline scripts.
-		# Should be removed once #21651 is fixed.
-		if( 'view.php' == basename( $_SERVER['SCRIPT_NAME'] ) ||
-			'bug_report_page.php' == basename( $_SERVER['SCRIPT_NAME'] ) ) {
-			http_csp_add( 'script-src', "'unsafe-inline'" );
-		}
-
 		# The JS Calendar control does unsafe eval, remove once we upgrade the control (see #20040)
 		if( 'bug_update_page.php' == basename( $_SERVER['SCRIPT_NAME'] ) ) {
 			http_csp_add( 'script-src', "'unsafe-eval'" );
